requirements.txt
py-multibase>=1.0.3
